+++ internal/subscriber/repo.go
@@ -8,27 +8,32 @@ import (
 )
 
 type Repo struct {
-	db *gorm.DB
+	conn *gorm.DB
 }
 
-func NewRepo(db *gorm.DB) *Repo {
-	return &Repo{db: db}
+func NewRepo(conn *gorm.DB) *Repo {
+	return &Repo{conn: conn}
 }
 
 func (r *Repo) Create(item *Subscriber) error {
-	return r.db.Create(item).Error
+	return r.conn.Create(item).Error
 }
 
 func (r *Repo) Update(item *Subscriber) error {
-	return r.db.Save(item).Error
+	return r.conn.Save(item).Error
 }
 
 func (r *Repo) GetByID(id uuid.UUID) (*Subscriber, error) {
-	t := Subscriber{ID: id}
-	request := r.db.Take(&t)
-	if err := request.Error; err != nil {
+	var sub Subscriber
+
+	err := r.conn.
+		Where(Subscriber{ID: id}).
+		First(&sub).
+		Error
+
+	if err != nil {
 		return nil, fmt.Errorf("get subscriber by id #%s: %w", id, err)
 	}
 
-	return &t, nil
+	return &sub, nil
 }